requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pymongo==4.6.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
pyjwt==2.8.1
transformers==4.36.0
torch==2.2.0
huggingface-hub==0.21.0
requests==2.31.0
jinja2==3.1.2
typer==0.9.0
rich==13.7.0
aiofiles==23.2.1
python-multipart==0.0.6
corsheaders==4.3.1
email-validator==2.1.0
uuid==1.30